elt/requirements.txt
boto3==1.34.162
pandas==2.2.2
pyarrow==16.1.0
requests==2.32.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0

